delete_update_search_users_profile

+++ rest-api/app/Http/Controllers/Api/UserController.php
@@ -31,13 +31,26 @@ class UserController extends Controller
             'interests.*' => 'string',
         ]);
 
+        // $user->update([
+        //     ...$dataToUpdate,
+        //     'interests' => isset($dataToUpdate['interests']) ? json_encode($dataToUpdate['interests']) : $user->interests,
+        // ]);
         $user->update([
-            ...$dataToUpdate,
-            'interests' => isset($dataToUpdate['interests']) ? json_encode($dataToUpdate['interests']) : $user->interests,
+            ...$dataToUpdate
         ]);
+
 
         return new UserResource($user);
     }
+
+    public function destroyOwnProfile(Request $request)
+    {
+        $user = $request->user();
+        $user->delete();
+
+        return response()->json(['message' => 'Your account has been deleted.']);
+    }
+
 
     // Show 10 users for admin
     public function index()
@@ -47,6 +60,35 @@ class UserController extends Controller
 
         return UserResource::collection($users);
     }
+    //Update as admin
+    public function adminUpdate(Request $request, $id)
+    {
+        $this->authorizeAdmin();
+
+        $user = User::findOrFail($id);
+
+        $dataToUpdate = $request->validate([
+            'name' => 'sometimes|string|max:255',
+            'bio' => 'nullable|string',
+            'location' => 'nullable|string|max:255',
+            'interests' => 'nullable|array',
+            'interests.*' => 'string',
+            'role' => ['sometimes', Rule::in(['USER', 'ADMIN'])], // allow to change the role
+        ]);
+
+        // $user->update([
+        //     ...$dataToUpdate,
+        //     'interests' => isset($dataToUpdate['interests']) ? json_encode($dataToUpdate['interests']) : $user->interests,
+        // ]);
+
+        $user->update([
+            ...$dataToUpdate
+        ]);
+
+
+        return new UserResource($user);
+    }
+
 
     // Delete  user as admin
     public function destroy($id)
@@ -54,10 +96,31 @@ class UserController extends Controller
         $this->authorizeAdmin();
 
         $user = User::findOrFail($id);
+
+        if ($user->id === auth()->id()) {
+            return response()->json(['error' => 'Admins cannot delete themselves'], 403);
+        }
         $user->delete();
 
         return response()->json(['message' => 'User deleted']);
     }
+
+    public function search(Request $request)
+    {
+        $this->authorizeAdmin();
+
+        $query = strtolower($request->input('query'));
+
+        $users = User::whereRaw('LOWER(name) LIKE ?', ["%{$query}%"])
+            ->orWhereRaw('LOWER(email) LIKE ?', ["%{$query}%"])
+            ->orWhereRaw('LOWER(location) LIKE ?', ["%{$query}%"])
+            ->with(['reviews', 'comments'])
+            ->paginate(10);
+
+        return UserResource::collection($users);
+    }
+
+
 
     // Is admin validation
     protected function authorizeAdmin()
